Move JS handler interface to separate file
This is how the Python and Go SDKs do it.
